initial set up for themeing and variables
