+++ src/styles/matic.css
@@ -5,7 +5,7 @@
 }
 
 body {
-  @apply bg-background font-sans text-foreground antialiased;
+  @apply bg-background text-foreground font-sans antialiased;
 }
 
 /* Utilities */
@@ -53,7 +53,7 @@ h6,
 
 p,
 .p {
-  @apply text-base leading-7 text-foreground;
+  @apply text-foreground text-base leading-7;
 }
 
 /* Blockquotes */
@@ -97,7 +97,7 @@ hr {
 /* Links */
 
 a {
-  @apply underline-offset-2 transition-all hover:text-muted-foreground;
+  @apply hover:text-muted-foreground underline-offset-2 transition-all;
 }
 
 /* Strong */
@@ -109,11 +109,11 @@ strong {
 /* Code */
 
 code {
-  @apply relative rounded bg-muted px-[0.3rem] py-[0.2rem] font-mono text-sm font-semibold;
+  @apply bg-muted relative rounded px-[0.3rem] py-[0.2rem] font-mono text-sm font-semibold;
 }
 
 pre {
-  @apply overflow-x-auto rounded-lg border bg-muted p-4;
+  @apply bg-muted overflow-x-auto rounded-lg border p-4;
 
   code {
     @apply relative rounded bg-transparent p-0;
@@ -147,7 +147,7 @@ table {
     }
 
     td {
-      @apply border px-4 py-2 text-left text-muted-foreground [&[align=center]]:text-center [&[align=right]]:text-right;
+      @apply text-muted-foreground border px-4 py-2 text-left [&[align=center]]:text-center [&[align=right]]:text-right;
     }
   }
 }
@@ -156,7 +156,7 @@ table {
 
 figure {
   figcaption {
-    @apply mt-2 text-xs text-muted-foreground;
+    @apply text-muted-foreground mt-2 text-xs;
   }
 }
 
@@ -175,7 +175,7 @@ video {
 /* Small */
 
 small {
-  @apply text-sm font-medium leading-none;
+  @apply text-sm leading-none font-medium;
 }
 
 /* Subscript and Superscript */
@@ -204,7 +204,7 @@ details {
 /* Keyboard */
 
 kbd {
-  @apply rounded-md border bg-muted px-1.5 py-0.5 font-mono text-sm;
+  @apply bg-muted rounded-md border px-1.5 py-0.5 font-mono text-sm;
 }
 
 /* Mark */
@@ -222,7 +222,7 @@ mark {
 /* Lead Paragraph */
 
 .lead {
-  @apply text-xl text-muted-foreground;
+  @apply text-muted-foreground text-xl;
 }
 
 /* Large Text */
@@ -234,13 +234,13 @@ mark {
 /* Small Text */
 
 .small {
-  @apply text-sm font-medium leading-none;
+  @apply text-sm leading-none font-medium;
 }
 
 /* Muted Text */
 
 .muted {
-  @apply text-sm text-muted-foreground;
+  @apply text-muted-foreground text-sm;
 }
 
 /* Lists within Lists */
@@ -268,19 +268,19 @@ dt {
 }
 
 dd {
-  @apply ml-4 text-muted-foreground;
+  @apply text-muted-foreground ml-4;
 }
 
 /* Blockquote Citations */
 
 blockquote cite {
-  @apply block text-sm text-muted-foreground;
+  @apply text-muted-foreground block text-sm;
 }
 
 /* Code Block Filename */
 
 pre[data-filename]::before {
-  @apply block text-sm font-medium text-muted-foreground;
+  @apply text-muted-foreground block text-sm font-medium;
   content: attr(data-filename);
 }
 
@@ -301,7 +301,7 @@ pre[data-filename]::before {
 /* Footnotes */
 
 .footnotes {
-  @apply border-t pt-6 text-sm text-muted-foreground;
+  @apply text-muted-foreground border-t pt-6 text-sm;
 }
 
 /* Task Lists */
@@ -328,7 +328,7 @@ button {
   }
 
   &:focus-visible {
-    @apply outline-none ring-2 ring-primary ring-offset-2;
+    @apply ring-primary ring-2 ring-offset-2 outline-none;
   }
 
   &:active {
@@ -348,7 +348,7 @@ select {
   }
 
   &:focus {
-    @apply border-primary outline-none ring-2 ring-primary ring-offset-2;
+    @apply border-primary ring-primary ring-2 ring-offset-2 outline-none;
   }
 }
 
@@ -431,21 +431,21 @@ select {
 
   /* Headings */
   h1 {
-    @apply mb-8 mt-6;
+    @apply mt-6 mb-8;
   }
 
   h2 {
-    @apply mb-6 mt-12;
+    @apply mt-12 mb-6;
   }
 
   h3 {
-    @apply mb-4 mt-8;
+    @apply mt-8 mb-4;
   }
 
   h4,
   h5,
   h6 {
-    @apply mb-4 mt-6;
+    @apply mt-6 mb-4;
   }
 
   /* Footnotes */
